+++ utils/energymon-overhead.c
@@ -7,6 +7,7 @@
  */
 #define _GNU_SOURCE
 #include <errno.h>
+#include <getopt.h>
 #include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -14,13 +15,42 @@
 #include "energymon-get.h"
 #include "energymon-time-util.h"
 
-int main(void) {
+static const char short_options[] = "+h";
+static const struct option long_options[] = {
+  {"help",      no_argument,       NULL, 'h'},
+  {0, 0, 0, 0}
+};
+
+static void print_usage(int exit_code) {
+  fprintf(exit_code ? stderr : stdout,
+          "Usage: energymon-overhead [OPTION]...\n\n"
+          "Measure the overhead of the init, read, and finish functions. Results are in\n"
+          "nanoseconds.\n\n"
+          "Note that overhead readings can only be as precise as the system clock supports.\n\n"
+          "Options:\n"
+          "  -h, --help               Print this message and exit\n");
+  exit(exit_code);
+}
+
+int main(int argc, char** argv) {
   char source[64] = { 0 };
   static energymon em;
   uint64_t time_start_ns, time_end_ns;
   uint64_t finit_ns, fread_ns, ffinish_ns;
   uint64_t energy_uj;
   int ret;
+  int c;
+  while ((c = getopt_long(argc, argv, short_options, long_options, NULL)) != -1) {
+    switch (c) {
+      case 'h':
+        print_usage(0);
+        break;
+      case '?':
+      default:
+        print_usage(1);
+        break;
+    }
+  }
 
   // no need to profile getting the energymon
   if (energymon_get(&em)) {